Add support for VisualEditor
This change makes a few changes to the skin that allow VisualEditor to
work with the skin. #mw_content has been renamed to #content,
the element #f-list has been renamed to #footer-info, the ids of the
children of #footer-info are now prefixed with footer-info-. The first
change is required for VisualEditor to run and the second two Changes
allow VisualEditor to update the last edit time. Changes have also been
made to main.css to allow OOUI popup to have a reasonable font size and
for the VisualEditor toolbar and save prompt to work as expected.

Bug: T177243
Change-Id: If5242f1930dde09c5520cc9aa5a692458a3f210e